fix: removing return types from thorest
