+++ packages/sdk/src/thor/thor-client/model/gas/EstimateGasOptions.ts
@@ -1,0 +1,18 @@
+import { Revision } from "@vcdm";
+
+/**
+ * Options for gas estimation.
+ */
+interface EstimateGasOptions {
+    /**
+     * The revision (block) to estimate gas at.
+     */
+    revision?: Revision;
+
+    /**
+     * Gas padding percentage to add on top of estimated gas (0, 1].
+     */
+    gasPadding?: number;
+}
+
+export type { EstimateGasOptions };
